modules/locationfeeder/requirements.txt
grpcio == 1.50.0
grpcio-tools == 1.50.0
kafka-python == 2.0.2
